refactor(BA-6668): move auth enums down from models to data
Move LoginSessionStatus/LoginAttemptResult into
data/auth/login_session_types.py (deleting models/login_session/enums.py)
and redirect data/auth's UserRole/UserStatus imports to their existing
homes in common.data.user.types and manager.data.user.types. This makes
data/auth a pure leaf and removes 2 more data->models back-edges feeding
the models<->data cycle. All non-data importers of the login-session
enums now point at the data layer directly.
