+++ src/ai/backend/manager/data/auth/login_session_types.py
@@ -1,10 +1,30 @@
 from __future__ import annotations
 
+import enum
 import uuid
 from dataclasses import dataclass
 from datetime import datetime
 
-from ai.backend.manager.models.login_session.enums import LoginAttemptResult, LoginSessionStatus
+
+class LoginSessionStatus(enum.StrEnum):
+    ACTIVE = "active"
+    INVALIDATED = "invalidated"
+    REVOKED = "revoked"
+
+
+class LoginAttemptResult(enum.StrEnum):
+    SUCCESS = "success"
+    FAILED_INVALID_CREDENTIALS = "failed_invalid_credentials"
+    FAILED_USER_INACTIVE = "failed_user_inactive"
+    FAILED_BLOCKED = "failed_blocked"
+    FAILED_PASSWORD_EXPIRED = "failed_password_expired"
+    FAILED_REJECTED_BY_HOOK = "failed_rejected_by_hook"
+    FAILED_SESSION_ALREADY_EXISTS = "failed_session_already_exists"
+    LOGOUT = "logout"
+    REVOKED_BY_ADMIN = "revoked_by_admin"
+    REVOKED_BY_USER = "revoked_by_user"
+    EVICTED = "evicted"
+    EXPIRED = "expired"
 
 
 @dataclass(frozen=True)
