+++ src/ai/backend/manager/repositories/auth/options.py
@@ -7,9 +7,12 @@ from typing import TYPE_CHECKING
 
 import sqlalchemy as sa
 
+from ai.backend.manager.data.auth.login_session_types import (
+    LoginAttemptResult,
+    LoginSessionStatus,
+)
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.condition_utils import make_string_in_factory
-from ai.backend.manager.models.login_session.enums import LoginAttemptResult, LoginSessionStatus
 from ai.backend.manager.models.login_session.row import LoginHistoryRow, LoginSessionRow
 
 if TYPE_CHECKING:
